python-usage-example/requirements.txt
huggingface-hub
orch==2.2.1
transformers==4.40.2